refactor: folder name
change folder name for more clarity
