+++ lambda/app/src/utils/helpers.ts
